fix: clamp & wrap SyncedAccountHeaderView at accessibility Dynamic Type sizes
The 42-char wallet address (deliberately never truncated, anti
vanity-spoofing) overflowed with no fallback at large Dynamic Type
sizes, and the status row packed four elements into one HStack that
clipped/overlapped at accessibility sizes.

- Address: clamp font growth to `.accessibility2` and `.fixedSize`
  force-wrap so it wraps to multiple lines instead of clipping; never
  truncates.
- Status row: `ViewThatFits` falls back from a single-line HStack to a
  two-row layout when the four elements no longer fit; leading/trailing
  groups extracted as shared @ViewBuilder helpers.
- Match lastSynced to the context label's `.caption`; add an
  accessibilityHint for the disabled (empty-address) copy button.
- Relocate view helpers to a trailing extension; split platform
  defaults and previews into sibling files (file_length).
- Add xSmall / accessibility5 / iPhone-width preview variants proving
  the acceptance range.

Fixes #932
